stt/requirements.txt
absl-py==0.12.0
google-cloud-speech==2.3.0
grpcio==1.42.0
grpcio-tools==1.42.0
librosa==0.8.0
numba==0.53.1
numpy==1.20.2
protobuf==3.15.8
